ADD :: delete comment 캡슐화
